+++ src/exes/moc.ml
@@ -41,7 +41,9 @@ let argspec = Arg.align [
 
   "-v", Arg.Set Flags.verbose, " verbose output";
   "-p", Arg.Set_int Flags.print_depth, " set print depth";
-  "--hide-warnings", Arg.Clear Flags.print_warnings, " hide warnings"; ]
+  "--hide-warnings", Arg.Clear Flags.print_warnings, " hide warnings";
+  "-Werror", Arg.Set Flags.warnings_are_errors, " treat warnings as werrors";
+  ]
 
   @ Args.error_args
 
@@ -192,6 +194,12 @@ let () =
   Arg.parse_expand argspec add_arg usage;
   if !mode = Default then mode := (if !args = [] then Interact else Compile);
   Flags.compiled := (!mode = Compile || !mode = Idl);
+
+  if !Flags.warnings_are_errors && (not !Flags.print_warnings)
+  then begin
+    eprintf "moc: --hide-warnings and -Werror together do not make sense"; exit 1
+  end;
+
   process_profiler_flags ();
   try
     process_files !args