+++ src/mo_config/flags.ml
@@ -7,6 +7,7 @@ type compile_mode = WasmMode | ICMode | RefMode | WASIMode
 let trace = ref false
 let verbose = ref false
 let print_warnings = ref true
+let warnings_are_errors = ref false
 let print_depth = ref 2
 let release_mode = ref false
 let compile_mode = ref ICMode